refactor : 검수 대기 목록 스키마 shared로 분리
* 검수 대기 목록 스키마 4종(상태·항목·페이지네이션·목록)을 partner/_shared/model로 이동
* review 세그먼트의 스키마·타입·조회 함수 참조를 shared 경로로 변경
* review.schema에는 검수 현황 요약 스키마만 잔류
