+++ apps/web/src/app/partner/review/_api/get-review-list.ts
@@ -1,7 +1,7 @@
 import { API_BASE_URL } from "@/shared/api/config";
 import { fetchJson } from "@/shared/api/fetch-json";
 import type { ReviewListFilter } from "@/shared/api/query-keys";
-import { reviewListSchema } from "../_model/review.schema";
+import { reviewListSchema } from "../../_shared/model/review-list.schema";
 import type { ReviewList } from "../_model/types";
 
 export function getReviewList(filter?: ReviewListFilter): Promise<ReviewList> {
